In htmlmodify.l, the tag-attribute pattern feature has been extended with the capability to pattern match the script text of a script element. The pattern for the script text can be specified using a text pseudo attribute in the tag-attribute pattern.
In configmisc.c, a variant of the strcasestr() function was added that
allows the length of the haystack to be specified, so that it doesn't
necessarily have to be null terminated.

The make_bitvector.pl script was extended with the capability to add
line-control preprocessor directives to the output file, so that
compiler error messages point to the original source file, rather than
the derived one.

In miscencdec.c.pre, the HTMLString() function now accepts an optional
length parameter to allow input strings that are not necessarily null
terminated. A similar modification has also been made to
HTMLcommentstring().