Extend tag-attribute pattern sequences with the capability to include sibling context.
Syntactically, a tag-attribute pattern sequence may now be preceded by
a another pattern sequence followed by a semicolon. The target pattern
of the sequence before the semicolon must match the previous HTML
sibling element (i.e. a child of the same parent) in order for the
pattern behind the semicolon to potentially match the next
sibling. The target of the whole pattern sequence is the target of the
sequence behind the last semicolon.