In htmlmodify.l, implemented a better method for resolving targets with mismatching first-line descendant patterns using reference counts.
Also eliminated some remaining possibilities for matching errors.
A pattern which is a first-line descendant of a target needs to have
the associated state node removed (instead of its lastm value
decreased) when the pattern is popped of the pattern stack. The lastm
value of a state node should not be advanced more than once by the
same tag token. A pattern-options node should not be placed on a
pattern state list more than once.